refactor http handlers
